# docs/swagger.yaml
basePath: /
definitions:
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.Jadwal:
    properties:
      _id:
        type: string
      estimasi_tiba:
        type: string
      kendaraan_id:
        type: string
      rute_id:
        type: string
      tanggal:
        type: string
      waktu_berangkat:
        type: string
    type: object
...
  models.Rute:
    properties:
      _id:
        type: string
      asal:
        type: string
      jarak_km:
        type: integer
      kode_rute:
        type: string
      nama_rute:
        type: string
      tujuan:
        type: string
    type: object
  models.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  repository.JadwalWithRute:
    properties:
      estimasi_tiba:
        type: string
      id:
        type: string
      rute:
        $ref: '#/definitions/models.Rute'
      rute_id:
        type: string
      tanggal:
        type: string
      waktu_berangkat:
        type: string
    type: object
...
host: localhost:8088
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample server for a transport app.
  license:
    name: Apache 2.0
    url: https://github.com/Haekalss
  title: Transport App API
  version: "1.0"
paths:
  /api/jadwals:
    get:
      consumes:
      - application/json
      description: Mengambil semua data jadwal beserta detail rutenya
      produces:
      - application/json
      responses:
        "200":
          description: Daftar semua jadwal
          schema:
            items:
              $ref: '#/definitions/repository.JadwalWithRute'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all jadwal
      tags:
      - Jadwal
    post:
      consumes:
      - application/json
      description: Membuat data jadwal baru (Admin Only)
      parameters:
      - description: Data jadwal baru
        in: body
        name: jadwal
        required: true
        schema:
          $ref: '#/definitions/models.Jadwal'
      produces:
      - application/json
      responses:
        "201":
          description: Jadwal berhasil dibuat
          schema:
            $ref: '#/definitions/models.Jadwal'
        "400":
          description: Bad Request - data tidak valid
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Rute or Kendaraan not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new jadwal
      tags:
      - Jadwal
  /api/jadwals/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus data jadwal berdasarkan ID (Admin Only)
      parameters:
      - description: Jadwal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Jadwal berhasil dihapus
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a jadwal
      tags:
      - Jadwal
    get:
      consumes:
      - application/json
      description: Mengambil data jadwal berdasarkan ID
      parameters:
      - description: Jadwal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Data jadwal yang ditemukan
          schema:
            $ref: '#/definitions/models.Jadwal'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Jadwal not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a jadwal by ID
      tags:
      - Jadwal
    put:
      consumes:
      - application/json
      description: Memperbarui data jadwal yang ada (Admin Only)
      parameters:
      - description: Jadwal ID
        in: path
        name: id
        required: true
        type: string
      - description: Data jadwal yang akan diupdate
        in: body
        name: jadwal
        required: true
        schema:
          $ref: '#/definitions/models.Jadwal'
      produces:
      - application/json
      responses:
        "200":
          description: Data jadwal diupdate
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Invalid ID atau Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden - Admin access required
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Rute not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an existing jadwal
      tags:
      - Jadwal
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"